Improve handling of parameters and documentation.
Regarding tcase/maintcas.c, the parameter handling was completely
reorganized. Before, options were parsed on splitted functions,
with a lot of redundancy between then. Now everything is centered
on main() and the options are global variables. Later one, such
option handling can be reused for other modules (for instance, -D
is used for almost every binary of Proteum).

Some functions relied on statically allocated variables and returning
pointers to those variables. This can and should be avoided. When
possible, functions that worked in this way were rewritten to pass
parameters by reference, removing the requirement for statically
allocated variables.

The documentation of several functions was rather improved. This is
specially true with lib/disco.c and lib/gerais.c.