Fix issue with monta_nome() and rewrite separa_param().
The previous implementation of separam_param() was barely working.
It was relying on a fixed size 'argv' array of strings, which would
work just for small programs. We now dynamically allocated 'argv'.
Another solution would be to set a much larger array as argv, but
that would be a waste of memory.